Refactored table specs to be a typescript files, added tests, added readme
