+++ grr/server/grr_response_server/gui/static/osquery_raw_table_specs.4.5.1.ts
@@ -1,4 +1,4 @@
-[
+export const tableSpecs451 = [
   {
     "cacheable":false,
     "evented":false,
