+++ grr/server/grr_response_server/gui/ui/components/flow_args_form/osquery_query_helper/osquery_query_helper.ts
@@ -20,7 +20,7 @@ import {
   tableCategoryFromSpecs,
   TableCategory,
   TableCategoryWithMatchMap,
-} from './osquery_helper_interfaces';
+} from './osquery_helper_model';
 import { FuzzyMatcher, Match } from '@app/lib/fuzzy_matcher';
