+++ grr/server/grr_response_server/gui/ui/tsconfig.json
@@ -29,9 +29,7 @@
     ],
     "paths": {
       "@app/*": ["./*"]
-    },
-    "resolveJsonModule": true,
-    "esModuleInterop": true
+    }
   },
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
